[PACKAGES-MIGRATION] migrate helper
